requirements.txt
asttokens
backcall
backports.functools-lru-cache
comm
debugpy
decorator
executing
importlib-metadata
ipykernel
ipython
jedi
jupyter_client
jupyter_core
matplotlib-inline
nest-asyncio
packaging
parso
pexpect
pickleshare
platformdirs
prompt-toolkit
psutil
ptyprocess
pure-eval
pycryptodome==3.18.0
Pygments
python-dateutil
pyzmq
six
stack-data
tornado
traitlets
typing_extensions
wcwidth
zipp
